fix: refactor file merging into different functions and files to reduce main function complexity
