+++ cli/install/src/lib/dependencies.ts
@@ -4,8 +4,14 @@ import ora from "ora";
 
 const execPromise = promisify(exec);
 
-export async function installDependencies(manager: string, packageDir: string): Promise<void> {
-  // try to install dependencies with the manager specified by the user, else fallback to npm
+export async function installDependencies(
+  manager: string,
+  packageDir: string,
+  isFirst: boolean = false,
+): Promise<void> {
+  console.log("\x1b[90m│");
+  if (!isFirst) console.log("\x1b[90m│");
+
   const installCommand = `${manager} install`;
   const spinner = ora(`Installing dependencies using ${manager}`).start();
