+++ cli/install/src/lib/merge/concat.ts
@@ -1,0 +1,5 @@
+import { PatchFilesConcat } from "../../schemas/patchFiles.js";
+
+// TODO: Implement the following functions
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function applyConcatMerge(targetPath: string, patch: PatchFilesConcat): void {}
