+++ cli/install/src/lib/merge/json.ts
@@ -1,0 +1,5 @@
+import { PatchFilesJson } from "../../schemas/patchFiles.js";
+
+// TODO: Implement the following functions
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function applyJsonMerge(targetPath: string, patch: PatchFilesJson): void {}
